+++ common/fabricSrc/main/java/ninjaphenix/expandedstorage/block/entity/OldChestBlockEntity.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2021 NinjaPhenix
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package ninjaphenix.expandedstorage.block.entity;
 
 import net.minecraft.block.BlockState;
@@ -7,6 +22,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
+import ninjaphenix.expandedstorage.block.AbstractChestBlock;
 import ninjaphenix.expandedstorage.block.OpenableBlock;
 import ninjaphenix.expandedstorage.block.entity.extendable.InventoryBlockEntity;
 import ninjaphenix.expandedstorage.block.entity.extendable.OpenableBlockEntity;
@@ -33,8 +49,8 @@ public class OldChestBlockEntity extends InventoryBlockEntity {
     }
 
     @Override
-    protected boolean shouldStateUpdateInvalidateItemAccess(BlockState oldState, BlockState state) {
-        return false;
+    protected boolean shouldStateUpdateInvalidateItemAccess(BlockState oldState, BlockState newState) {
+        return oldState.get(AbstractChestBlock.CURSED_CHEST_TYPE) != newState.get(AbstractChestBlock.CURSED_CHEST_TYPE);
     }
 
     @Override