+++ changelog.md
@@ -3,3 +3,4 @@
 - Fixed some issues with Piglin guarding tag and also revert to the old tag format.
 - Fixed the map color of Christmas chest variants
 - Remove some unnecessary data from models as well as some minor relocation for barrel and old chest base models.
+- Major resource / data pack renames, generally moving from christmas chest -> present